move edit components to their own folder
